+++ lib/kward/tools/tool_call.rb
@@ -19,6 +19,7 @@ module Kward
       "code_search" => "code_search",
       "summarize_file_structure" => "summarize_file_structure",
       "context_for_task" => "context_for_task",
+      "context_budget_stats" => "context_budget_stats",
       "retrieve_tool_output" => "retrieve_tool_output",
       "web_search" => "web_search",
       "fetch_content" => "fetch_content",